Added a new service for string formating. Some bug-fixing done. Added documentation for components: - PublicationCard. - CommentBox. - TextAndImageForm Added documentation for services: - StringFormat
