+++ src/app/modules/app-components/README.md
@@ -21,6 +21,9 @@ Los datos que se deben incluir en la documentación de cada componente son:
 * [Single-Field Form n2options](#Single-field-Form-n2options)
 * [ElementCard](#ElementCard)
 * [ElementCardBox](#ElementCardBox)
+* [TextAndIMageForm](#TextAndIMageForm)
+* [CommentBox](#CommentBox)
+* [PublicationCard](#PublicationCard)
 
 ### Single-Field Form n2options.
 
@@ -68,4 +71,50 @@ los valores mediante las propiedades Input icon y name.
 
 **Selector**: `app-element-card-box`.
 
+### TextAndImageForm.
 
+**Descripción**: Formulario con una caja de texto de tamaño auto-ajustable al texto contenido con soporte para adjuntar una imágen.
+El envío del contenido se dispara mediante "enter" o el botón dedicado.
+
+**Entradas**:
+* `textInputLabel`: Label para la caja de texto. String.
+* `textInputPlaceholder`: Placeholder para la caja de texto. String.
+
+**Salidas**:
+* `newContent`: Evento que se dispara cuando el contenido del formulario se envía. Datos de retorno del evento son:
+  * `text`: Un `string` representando el contenido de la caja de texto.
+  * `image`: Un objeto de tipo `File` representando la imágen.
+
+**Selector**: `app-text-and-image-form`.
+
+### CommentBox.
+
+**Descripción**: Caja para desplegar comentarios y crear comentarios. Los comentarios soportan texto e imagen.
+
+**Entradas**:
+* `comments`: Arreglo de tipo `Comment` (revisar en las interfaces para publicaciones).
+
+**Salidas**:
+* `moreComments`: Evento que se dispara cuando se presiona el botón para recuperar más comentarios. No retorna datos.
+* `newComment`: Evento que se dispara cuando se envía un nuevo comentario. Datos de retorno:
+  * `text`: `string` representanto el texto del comentario.
+  * `image`: Objeto `File` representando la imágen adjuntada.
+
+**Selector**: `app-comment-box`.
+
+### PublicationCard.
+
+**Descripción**: Card para desplegar la información de una publicación.
+
+**Entradas**:
+`profileImgSrc`: `string`. La URL de la imagen de perfil del usuario autor de la publicación.
+`profileName`: `string`. Nombre completo del usuario autor.
+`publicationImgSrc`: `string`. URL de la imagen de la publicación.
+`text`: `string`. Texto de la publicación.
+`subtitle`: `string`. Texto debajo del nombre del usuario.
+
+**Salidas**:
+* `action`: Evento que informa que comando se ejecuto en la publicación. Datos retornados:
+  * `name`: `string`. El nombre del comando (`comment`|`share`|`favorite`)
+
+**Selector**: `app-comment-box`.
